<feat> [be] : 화상강의 생성, 조회, 수정기능 구현

See merge request s11-webmobile4-sub2/S11P12D210!92

+++ backend/src/main/java/com/yoga/backend/common/entity/LiveLectures.java
@@ -25,7 +25,7 @@ public class LiveLectures {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long liveId; // 강의 ID (Primary Key)
+    private Integer liveId; // 강의 ID (Primary Key)
 
     @Column(length = 30, nullable = false)
     private String liveTitle; // 강의 제목
